[ASM] Refactor Value and assembler
When not supporting floating point number (defined ASM_NOFLOAT in
config_libasm.h), it uses uint32_t to represent values.

When supporting floating point number (comment out ASM_NOFLOAT in
config_libasm.h), it uses uint64_t and double to represent
values. Although it keep external integer API as 32-bit integer for
compatibilty and optimization.

+++ test/test_asserter.h
@@ -46,9 +46,9 @@ struct TestAsserter {
     void isNull(const char *file, int line, const char *message, const void *actual);
     void isNotNull(const char *file, int line, const char *message, const void *actual);
 
-    /** assert uint32_t */
+    /** assert uint64_t */
     void equals(
-            const char *file, int line, const char *message, uint32_t expected, uint32_t actual);
+            const char *file, int line, const char *message, uint64_t expected, uint64_t actual);
 
     /** assert const char* */
     void equals(const char *file, int line, const char *message, const char *expected,
